refactor: rename migrations and update migration scripts
